Fixed the formating to match the example. Added a console.clear() in the main program.

+++ final/Foundation4/Swimming.cs
@@ -1,7 +1,7 @@
 public class Swimming : Activity
 {
-    protected int _laps;
-    public Swimming(string date, string name, double minutes, int laps) : base(date, name, minutes)
+    protected double _laps;
+    public Swimming(string date, string name, double minutes, double laps) : base(date, name, minutes)
     {
         _laps = laps;
     }
